requirements.txt
python-dotenv
flask
langchain
langchain-groq
langchain-community
youtube_transcript_api
pypdf